TweenGroups and 'removeById' event system method - Added 'removeById' method to the event system that removes a specific event listener by its unique ID. - Added the 'TweenGroup' object and the 'createGroup' method to the tweenManager. - Added the 'canPurge' property to the Tween object. - Updated the 'emit' method of the event method to attach the listener ID to the passed data. - Updated the 'emit' method of the event method to return the parent object instead of a boolean. - Removed ability to apply the 'outline' property to circular sprites. - Removed in-built shortening of the 'child' system namespace to 'c'.

+++ src/Game/tweenManager/index.js
@@ -1,8 +1,9 @@
 // MobSin.game.tweenManager
 
-module.exports = (_game) => {
-	let Tween = require("./Tween");
+let Tween = require("./Tween");
+let TweenGroup = require("./TweenGroup");
 
+module.exports = (_game) => {
 	_game.tweens = [];
 	_game.tweenManager = {
 		create: (_obj, from, to, time, presets) => {
@@ -11,6 +12,9 @@ module.exports = (_game) => {
 			) - 1;
 
 			return _game.tweens[newInd];
+		},
+		createGroup: (tweens, presets) => {
+			return new TweenGroup(tweens, presets);
 		},
 		getAll: () => {
 			return _game.tweens;
@@ -22,7 +26,10 @@ module.exports = (_game) => {
 			for (let i = _game.tweens.length - 1; i >= 0; i--) {
 				_game.tweens[i]._update();
 
-				if (_game.tweenManager.purge && _game.tweens[i].finished) {
+				if (_game.tweenManager.purge &&
+					_game.tweens[i].finished &&
+					_game.tweens[i].canPurge
+				) {
 					_game.object.destroyById(_game.tweens[i]._id);
 					_game.tweens.splice(i, 1);
 				}
